send shutdown to stdin instead of os signal(kill) on windows

+++ electron/js/server.js
@@ -7,6 +7,7 @@ const { app, dialog, shell } = require('electron');
 const Util = require('./util.js');
 
 let maxStdErrChunksBuffer = 10;
+const winShutdownStdinMessage = 'shutdown\n';
 
 class Server {
 
@@ -114,7 +115,12 @@ class Server {
 				});
 				
 				this.stopTriggered = true;
-				this.cp.kill(signal);
+ 				if (process.platform === 'win32') {
+					 // it is not possible to handle os signals on windows, so we can't do graceful shutdown on go side
+					this.cp.stdin.write(winShutdownStdinMessage);
+				} else {
+					this.cp.kill(signal);
+				}
 			} else {
 				resolve();
 			};
